+++ src/New/Entity/AttachEffectClass.cpp
@@ -635,6 +635,73 @@ bool AttachEffectClass::ShouldBeDiscardedNow()
 		return true;
 	}
 
+	if ((discardOn & DiscardCondition::Ammo) != DiscardCondition::None)
+	{
+		bool trigger = false;
+		if (pType->DiscardOn_Ammo_Min.isset() || pType->DiscardOn_Ammo_Max.isset())
+		{
+			const int min = pType->DiscardOn_Ammo_Min.Get(-1);
+			const int max = pType->DiscardOn_Ammo_Max.Get(-1);
+			const int ammo = pTechno->Ammo;
+
+			trigger = (min < 0 || ammo >= min) && (max < 0 || ammo <= max);
+		}
+
+		if (trigger)
+		{
+			this->LastDiscardCheckValue = true;
+			return true;
+		}
+	}
+
+	if ((discardOn & DiscardCondition::Health) != DiscardCondition::None)
+	{
+		if (auto const pTypeData = pTechno->GetTechnoType())
+		{
+			const double hp = pTechno->GetHealthPercentage();
+
+			if (pType->DiscardOn_Health_Min.isset() || pType->DiscardOn_Health_Max.isset())
+			{
+				const double min = pType->DiscardOn_Health_Min.Get(0.0);
+				const double max = pType->DiscardOn_Health_Max.Get(1.0);
+
+				if ((hp > 0.0 ? hp > min : hp >= min) && hp <= max)
+				{
+					this->LastDiscardCheckValue = true;
+					return true;
+				}
+			}
+		}
+	}
+
+	if ((discardOn & DiscardCondition::LandType) != DiscardCondition::None)
+	{
+		if (pType->DiscardOn_LandTypes != LandTypeFlags::None)
+		{
+			if (auto const pCell = pTechno->GetCell())
+			{
+				LandTypeFlags landFlags = pType->DiscardOn_LandTypes;
+				if (IsLandTypeInFlags(landFlags, pCell->LandType))
+				{
+					this->LastDiscardCheckValue = true;
+					return true;
+				}
+			}
+		}
+	}
+	
+	if ((discardOn & DiscardCondition::Mission) != DiscardCondition::None)
+	{
+		if (pType->DiscardOn_Missions.size() > 0)
+		{
+			if (pType->DiscardOn_Missions.Contains(pTechno->CurrentMission))
+			{
+				this->LastDiscardCheckValue = true;
+				return true;
+			}
+		}
+	}
+
 	if (pTechno->Target)
 	{
 		const bool inRange = (discardOn & DiscardCondition::InRange) != DiscardCondition::None;